AUTO-6119: Makefile: Add "upgrade-base-dependencies" target
Add a "upgrade-base-dependencies" target to Makefile which must be run
whenever the dependencies are changed. It builds the "builder"
Dockerfile target, specifying "true" for the GENERATE_REQUIREMENTS_TXT
build arg. This causes the ./requirement.txt file not to be used to
install the dependencies so that "pip freeze" can then be run in the
container to generate a new ./requirements.txt .

Document this in README.md .