AUTO-6119: Dockerfile: Use or generate requirements.txt
Add a GENERATE_REQUIREMENTS_TXT build arg which controls whether what's
in ./requirements.txt is installed. When it's "true", the unpinned (but
constrained) dependencies found in the requirements.txt from lgsvl
(PythonApi) are installed so that a new ./requirements.txt can be
generated.

Include a 0-sized ./requirements.txt so that the COPY doesn't fail
before a "real" one has been generated.