AUTO-6362: Support building with arbitrary versions of the Python base image
Have Makefile recognize a PYTHON_VERSION_TAG variable to be set on the
make command line. Use it to set build args added to docker/Dockerfile
that parameterize the Python version used by the image and to form the
tag of the image built by "make build-base" (but not the other images).

Because the dependency versions pinned by the requirements file will
vary based on the version of Python, generate and use a
requirements-PYTHON_VERSION_TAG.txt .

If PYTHON_VERSION_TAG is not specified, nothing is changed: the version
of Python specified by the defaults of the Dockerfile build args is
used, the image built by "make build-base" is tagged with "latest", and
the basename of the requirements file is not prefixed with
"-PYTHON_VERSION_TAG".